% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRexplorer.R
\name{PCRexplorer}
\alias{PCRexplorer}
\title{Visualisation of nucleic acid fragments size versus concentration profiles.}
\usage{
PCRexplorer(
  dat,
  targets,
  tolerance = NULL,
  threshold = NULL,
  logx = FALSE,
  logy = FALSE,
  xlimits = NULL,
  join = FALSE,
  control = NULL
)
}
\arguments{
\item{dat}{a data frame holding results of a PCR amplification and electrophoresis run.}

\item{targets}{a named numeric vector of diagnostic fragment sizes.}

\item{tolerance}{a numeric vector with the left and right tolerance values.}

\item{threshold}{a numeric scalar with the threshold concentration value.}

\item{logx}{logical operator to display fragments size values on logarithmic scale.}

\item{logy}{logical operator to display concentration values on logarithmic scale.}

\item{xlimits}{numeric vector restricting visualisation of fragments size range within min and max window.}

\item{join}{logical operator to show lines joining points from the same sample.}

\item{control}{string of the sample identifier name to show lines joining points of template samples.}
}
\description{
In order to make an informed choice for the tolerance and threshold settings, it is useful to visualise in a single plot the full range
of fragment sizes and concentrations returned by a PCR run, as well as their mutual relationships. For specific examples about how to
interpret and make use of this plot, see the package vignette.
}
\examples{
# load the data
data(mosquito)

# parameter settings
targets <- c(arabiensis = 315, gambiae = 390, melas = 464)
tolerance <- c(0,10)
threshold <- 0.05

PCRexplorer(mosquito, targets, tolerance, threshold, logx = TRUE, logy = TRUE, join = TRUE, control = "CTR")

}
