% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRoutcome.R
\name{PCRoutcome}
\alias{PCRoutcome}
\title{Classification of PCR samples based on diagnostic nucleic acid fragments.}
\usage{
PCRoutcome(dat, targets, tolerance, threshold)
}
\arguments{
\item{dat}{a data frame holding results of a PCR amplification and electrophoresis run.}

\item{targets}{a named numeric vector of diagnostic fragment sizes.}

\item{tolerance}{a numeric vector with the left and right tolerance values.}

\item{threshold}{a numeric scalar with the threshold concentration value.}
}
\value{
A data frame with the PCR plate well identifier, the sample identifier, and the classification outcome.
}
\description{
Takes an input data frame with the size/concentration profile of DNA/RNA fragments, and returns an output data frame of samples classified
based on the size of diagnostic fragments.
}
\details{
The \code{targets} argument must be a named vector, where names correspond to the classification categories (e.g. species, clones, etc.).
Note that based on the specific combination of \code{targets}, \code{tolerance}, and \code{threshold} parameters, a sample can be classified
in more than a single category. This may have biological significance or not at all, and should be checked a posteriori.
Samples that cannot be assigned to a category return NA.
}
\examples{
# load the data
data(mosquito)

# define a named vector of classification sizes (categories are species names in this case)
targets <-  c(arabiensis = 315, gambiae = 390, melas = 464)

# classify samples with a tolerance of ±10 bp and 0.2 ng/uL concentration
head(PCRoutcome(mosquito, targets, c(10,10), 0.2), head = 10L)

# the tolerance around the target size can be asymmetric
head(PCRoutcome(mosquito, targets, c(0,10), 0.2), head = 10L)

}
\seealso{
Other related functions: 
\code{\link{PCRpositive}()}
}
\concept{related functions}
