% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRpherogram.R
\name{PCRpherogram}
\alias{PCRpherogram}
\title{Visualise amplification results as pherograms in a PCR plate layout.}
\usage{
PCRpherogram(dat, target_size, tolerance, threshold)
}
\arguments{
\item{dat}{a data frame holding results of the PCR amplification and electrophoresis run.}

\item{target_size}{a numeric scalar of the diagnostic fragment size.}

\item{tolerance}{a numeric vector with the left and right tolerance values.}

\item{threshold}{a numeric scalar with the threshold concentration value.}
}
\description{
The plot mimics a collection of electropherograms for each of the samples in a PCR microwell plate, where 'bands' are represented as vertical bars
whose color depends on the classification outcome: bars matching the classification criteria, shown in the plot title, are highlighted
in red, whereas 'parasite bands' are shown in blue (#F8766D and #619CFF, respectively, \pkg{ggplot2} default colors). The sample identifier name is shown
on the top left corner of each 'pherogram'.
}
\examples{
# load the data
data(mosquito)

PCRpherogram(mosquito, target_size = 390, tolerance = c(0, 10), threshold = 0.05)

}
