% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRprofilR-package.R
\docType{package}
\name{PCRprofilR-package}
\alias{PCRprofilR}
\alias{PCRprofilR-package}
\title{PCRprofilR: Classification of Nucleic Acids Fragments}
\description{
Profiling of nucleic acids fragments following PCR amplification, from output files of high-throughput capillary elecrophoresis bioanalyzers.
}
\author{
\strong{Maintainer}: Carlo Costantini \email{carlo.costantini@ird.fr} (\href{https://orcid.org/0000-0003-1016-129X}{ORCID})

}
\keyword{internal}
