% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRprofilR-package.R
\docType{package}
\name{PCRprofilR-package}
\alias{PCRprofilR}
\alias{PCRprofilR-package}
\title{PCRprofilR: Classification of Nucleic Acid Fragments}
\description{
Profiling nucleic acid fragments following PCR amplification, using output files from high-throughput capillary electrophoresis bioanalyzers.
}
\author{
\strong{Maintainer}: Carlo Costantini \email{carlo.costantini@ird.fr} (\href{https://orcid.org/0000-0003-1016-129X}{ORCID})

}
\keyword{internal}
